Major index.js refactor
