+++ index.js
@@ -1,20 +1,9 @@
-const fs = require('fs')
-
-const categoriesRepository = require('./core/knowledgeRepository/categoriesRepository')
-const notesRepository = require('./core/knowledgeRepository/notesRepository')
+const bootstrapRepositories = require('./bootstrapRepositories')
+const saveRepositories = require('./saveRepositories')
 const mapArgvToAction = require('./core/mapArgvToAction')
 const sendToRouter = require('./core/sendToRouter')
 
-const CATEGORIES_REPOSITORY_PATH = './categories.json'
-const NOTES_REPOSITORY_PATH = './notes.json'
-
-//CLI HANDLING
-const categories = loadDataFromRepository(CATEGORIES_REPOSITORY_PATH)
-const notes = loadDataFromRepository(NOTES_REPOSITORY_PATH)
-
-categories.forEach(categoriesRepository.add.bind(categoriesRepository))
-notes.forEach(notesRepository.add.bind(categoriesRepository))
-
+bootstrapRepositories()
 const actionToBePerformed = mapArgvToAction(process.argv)
 
 sendToRouter(actionToBePerformed)
@@ -22,25 +11,4 @@ sendToRouter(actionToBePerformed)
     .catch(err => console.error(`Operation failed. More info:\n${err}`))
     .then(process.exit)
 
-//LOCAL FUNCTIONS
-function handleRepositoryLoadingError(err) {
-    const MODULE_NOT_FOUND_ERROR = 'MODULE_NOT_FOUND'
-    if (err.code === MODULE_NOT_FOUND_ERROR) {
-        console.warn('Unable to find categories repository, creating new one')
-        return []
-    }
-    throw `Unhandled error during accessing categories repository: ${err.code}`
-}
-
-function loadDataFromRepository(path) {
-    try {
-        return require(path)
-    } catch (err) {
-        return handleRepositoryLoadingError(err)
-    }
-}
-
-function saveRepositories() {
-    fs.writeFileSync(CATEGORIES_REPOSITORY_PATH, JSON.stringify(categoriesRepository.getAll()))
-    fs.writeFileSync(NOTES_REPOSITORY_PATH, JSON.stringify(notesRepository.getAll()))
-}
+saveRepositories()
