+++ repositoriesPaths.js
@@ -1,0 +1,6 @@
+const path = require('path')
+
+module.exports = {
+    CATEGORIES: path.resolve('./categories.json'),
+    NOTES: path.resolve('./notes.json')
+}
